100% working libft

+++ ft_putstr.c
@@ -6,7 +6,7 @@
 /*   By: jhouston <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 08:55:08 by jhouston          #+#    #+#             */
-/*   Updated: 2019/06/05 16:32:51 by jhouston         ###   ########.fr       */
+/*   Updated: 2019/06/24 16:08:53 by jhouston         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (str == NULL)
